Out of server updates.
Running the updater process within the server blocks everything else
what with the lack of threads. This commit allows you to specify a url
to the server. It'll fetch the url of the next feed to be updated, do
the difficult work then send a json encoded feedparser result back to
the server process to be entered into the database.